reworking loans

+++ src/epoch/EpochExternal.sol
@@ -7,7 +7,7 @@ import {IEpochExternal} from '../interfaces/INuggFT.sol';
 import {EpochView} from '../epoch/EpochView.sol';
 
 abstract contract EpochExternal is IEpochExternal {
-    function epoch() external view override returns (uint256) {
+    function epoch() external view override returns (uint32) {
         return EpochView.activeEpoch();
     }
 }
